[refactor] 정산 기능 패키지 이름 변경
