+++ backend/src/main/java/com/fiiiiive/zippop/payout/repository/PayoutRepository.java
@@ -1,6 +1,6 @@
-package com.fiiiiive.zippop.settlement.repository;
+package com.fiiiiive.zippop.payout.repository;
 
-import com.fiiiiive.zippop.settlement.entity.Settlement;
+import com.fiiiiive.zippop.payout.entity.Payout;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SettlementRepository extends JpaRepository<Settlement, Long> {
+public interface SettlementRepository extends JpaRepository<Payout, Long> {
 
     @Query("SELECT s FROM Settlement s WHERE s.storeIdx = :storeIdx")
-    Optional<Page<Settlement>> findAllByStoreIdx(Long storeIdx, Pageable pageable);
+    Optional<Page<Payout>> findAllByStoreIdx(Long storeIdx, Pageable pageable);
 
 }
